fixup(cycle-005 PR-A3.9): iter-2 — format parity + legacy Rust preserved + harness rebuild + camelToKebab rune-safe
iter-1 bridge findings absorbed (verdict REQUEST_CHANGES; 0 BLOCKER, 0
HIGH-consensus, 4 PRAISE callouts; 7 medium findings, all actionable):

go-runner-lacks-format-checker-detail / rust-date-time-leap-second
(medium / iter-1 cluster — three-implementation date-time divergence
risk):
  Three independent date-time validators (Python regex, Rust hand-
  rolled byte-parser, Go santhosh-tekuri library default) created a
  drift surface — divergence on `+00:00` offset / fractional-second
  precision was a matter of when, not if. Resolved by registering a
  custom date-time format checker in Go matching the same regex
  shape Python and Rust use:

    `^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d{1,9})?Z$`

  Go's compiler.RegisterFormat takes a Format struct with Validate
  callback returning error. The custom checker replaces the
  library default for the date-time keyword while AssertFormat()
  ensures the format is treated as ASSERTIVE per cycle-005
  contracts. All three runners now share the same regex shape.

harness-cargo-build-path-mismatch (medium / iter-1):
  Bash harness's `if [[ ! -x "$RUST_BIN" ]]; then cargo build` path
  silently ran a stale binary after main.rs edits. Resolved by
  unconditionally invoking `cargo build --release --quiet` before
  every harness run. Cargo's incremental compilation makes the
  unconditional rebuild cheap (≤1 s when source unchanged); the
  silent-stale-binary failure mode is exactly what parity tests
  exist to prevent.

rust-main-rs-replaced-no-glob-runner (medium / iter-1):
  PR-A3.9 baseline replaced the legacy Rust runner (which walked
  domain-event / conversation / billing-entry / transfer-spec /
  lifecycle-transition-payload / health-status / thinking-trace
  multi-vector files) with the cross-runner. Coverage of v0/v1/v2
  schema corpus would have silently dropped. Restored as a
  separate binary:

    Cargo.toml [[bin]] name = "legacy-vectors", path = "src/main.rs"
    Cargo.toml [[bin]] name = "cross-runner",   path = "src/bin/cross-runner.rs"

  `glob` dependency restored. `cargo run` (default binary)
  continues to exercise the legacy corpus; `cargo run --bin
  cross-runner` runs the new harness emitter. No coverage regression.

F-001 (medium / iter-1):
  camelToKebab in Go used `for i, r := range s; s[i-1]` — byte
  index against a rune iterator. Latent trap on multibyte schema
  names. Resolved by switching to `runes := []rune(s); runes[i-1]`
  (rune-indexed throughout). ASCII names unchanged in output;
  Unicode-bearing names now correctly handled.

constraint-level-invalids-additions (medium / iter-1):
  OracleDigest's telegram-variant-multibyte-bypass entry needed
  rationale. Added `rationale_per_fixture` map to
  `vectors/_meta/constraint-level-invalids.json` documenting the
  LOCAL `utf8_byte_length_max` builtin (NFR-4 4 KB cap) targeting
  + cross-runner manifest semantics ('pass-constraint-level').
  Future v8.6.0 / v8.7.0 fixture additions land their rationale
  in the same map.

Persistent locked-de-scope (per Vision 008 / 009; documented in PR
body):
  harness-not-in-ci (medium): the cross-runner harness is opt-in
  via `npm run vectors:cross-runners`; NOT wired into `check:all`
  because Go/Rust toolchains are optional. CI consumers requiring
  full parity install all four toolchains; CI consumers running
  only `check:all` need only Node.js. The TS reference's
  `vectors:cross-runner` (singular) is in CI by virtue of being
  a vitest-exercised path. Adding a `.github/workflows/cross-
  runner.yml` for the multi-runtime harness is a v8.7.0 / FR-A2
  follow-up scope decision (toolchain-availability matrix per CI
  runner).

Net deltas this iteration:
- Custom date-time format checker in Go (rfc3339UTCDateTime regex
  matching Python/Rust shape exactly)
- camelToKebab rune-indexed for Unicode safety
- Always-rebuild Rust binary in harness (--quiet for log noise
  reduction)
- Legacy Rust runner restored as `legacy-vectors` binary;
  cross-runner moved to `src/bin/cross-runner.rs`
- glob dependency restored to Cargo.toml
- Rationale map added to constraint-level-invalids registry
- +0 schemas, +0 fixtures, +0 EVALUATOR_BUILTINS, +0 constraint files
- Tests still 8634/8634 pass; harness still emits 286 byte-identical
  manifest entries across TS / Python / Go / Rust
